requirements.txt
disnake
aiohttp
python-dotenv
